chore: add ssr benchmark suite
