+++ benchmarking/benchmarks/reactivity/kairo/kairo_broad.js
@@ -1,20 +1,25 @@
-import { assert, fastest_test } from '../../utils.js';
-import * as $ from '../../../packages/svelte/src/internal/client/index.js';
-import { busy } from './util.js';
+import { assert, fastest_test } from '../../../utils.js';
+import * as $ from 'svelte/internal/client';
 
 function setup() {
 	let head = $.state(0);
-	let computed1 = $.derived(() => $.get(head));
-	let computed2 = $.derived(() => ($.get(computed1), 0));
-	let computed3 = $.derived(() => (busy(), $.get(computed2) + 1)); // heavy computation
-	let computed4 = $.derived(() => $.get(computed3) + 2);
-	let computed5 = $.derived(() => $.get(computed4) + 3);
+	let last = head;
+	let counter = 0;
 
 	const destroy = $.effect_root(() => {
-		$.render_effect(() => {
-			$.get(computed5);
-			busy(); // heavy side effect
-		});
+		for (let i = 0; i < 50; i++) {
+			let current = $.derived(() => {
+				return $.get(head) + i;
+			});
+			let current2 = $.derived(() => {
+				return $.get(current) + 1;
+			});
+			$.render_effect(() => {
+				$.get(current2);
+				counter++;
+			});
+			last = current2;
+		}
 	});
 
 	return {
@@ -23,18 +28,19 @@ function setup() {
 			$.flush_sync(() => {
 				$.set(head, 1);
 			});
-			assert($.get(computed5) === 6);
-			for (let i = 0; i < 1000; i++) {
+			counter = 0;
+			for (let i = 0; i < 50; i++) {
 				$.flush_sync(() => {
 					$.set(head, i);
 				});
-				assert($.get(computed5) === 6);
+				assert($.get(last) === i + 50);
 			}
+			assert(counter === 50 * 50);
 		}
 	};
 }
 
-export async function kairo_avoidable_unowned() {
+export async function kairo_broad_unowned() {
 	// Do 10 loops to warm up JIT
 	for (let i = 0; i < 10; i++) {
 		const { run, destroy } = setup();
@@ -53,13 +59,13 @@ export async function kairo_avoidable_unowned() {
 	destroy();
 
 	return {
-		benchmark: 'kairo_avoidable_unowned',
+		benchmark: 'kairo_broad_unowned',
 		time: timing.time.toFixed(2),
 		gc_time: timing.gc_time.toFixed(2)
 	};
 }
 
-export async function kairo_avoidable_owned() {
+export async function kairo_broad_owned() {
 	let run, destroy;
 
 	const destroy_owned = $.effect_root(() => {
@@ -84,7 +90,7 @@ export async function kairo_avoidable_owned() {
 	destroy_owned();
 
 	return {
-		benchmark: 'kairo_avoidable_owned',
+		benchmark: 'kairo_broad_owned',
 		time: timing.time.toFixed(2),
 		gc_time: timing.gc_time.toFixed(2)
 	};
