+++ benchmarking/benchmarks/reactivity/kairo/kairo_repeated.js
@@ -1,5 +1,5 @@
-import { assert, fastest_test } from '../../utils.js';
-import * as $ from '../../../packages/svelte/src/internal/client/index.js';
+import { assert, fastest_test } from '../../../utils.js';
+import * as $ from 'svelte/internal/client';
 
 let size = 30;
