+++ benchmarking/benchmarks/reactivity/kairo/kairo_triangle.js
@@ -1,23 +1,35 @@
-import { assert, fastest_test } from '../../utils.js';
-import * as $ from '../../../packages/svelte/src/internal/client/index.js';
+import { assert, fastest_test } from '../../../utils.js';
+import * as $ from 'svelte/internal/client';
 
-let len = 50;
-const iter = 50;
+let width = 10;
+
+function count(number) {
+	return new Array(number)
+		.fill(0)
+		.map((_, i) => i + 1)
+		.reduce((x, y) => x + y, 0);
+}
 
 function setup() {
 	let head = $.state(0);
 	let current = head;
-	for (let i = 0; i < len; i++) {
+	let list = [];
+	for (let i = 0; i < width; i++) {
 		let c = current;
+		list.push(current);
 		current = $.derived(() => {
 			return $.get(c) + 1;
 		});
 	}
+	let sum = $.derived(() => {
+		return list.map((x) => $.get(x)).reduce((a, b) => a + b, 0);
+	});
+
 	let counter = 0;
 
 	const destroy = $.effect_root(() => {
 		$.render_effect(() => {
-			$.get(current);
+			$.get(sum);
 			counter++;
 		});
 	});
@@ -25,22 +37,24 @@ function setup() {
 	return {
 		destroy,
 		run() {
+			const constant = count(width);
 			$.flush_sync(() => {
 				$.set(head, 1);
 			});
+			assert($.get(sum) === constant);
 			counter = 0;
-			for (let i = 0; i < iter; i++) {
+			for (let i = 0; i < 100; i++) {
 				$.flush_sync(() => {
 					$.set(head, i);
 				});
-				assert($.get(current) === len + i);
+				assert($.get(sum) === constant - width + i * width);
 			}
-			assert(counter === iter);
+			assert(counter === 100);
 		}
 	};
 }
 
-export async function kairo_deep_unowned() {
+export async function kairo_triangle_unowned() {
 	// Do 10 loops to warm up JIT
 	for (let i = 0; i < 10; i++) {
 		const { run, destroy } = setup();
@@ -59,13 +73,13 @@ export async function kairo_deep_unowned() {
 	destroy();
 
 	return {
-		benchmark: 'kairo_deep_unowned',
+		benchmark: 'kairo_triangle_unowned',
 		time: timing.time.toFixed(2),
 		gc_time: timing.gc_time.toFixed(2)
 	};
 }
 
-export async function kairo_deep_owned() {
+export async function kairo_triangle_owned() {
 	let run, destroy;
 
 	const destroy_owned = $.effect_root(() => {
@@ -90,7 +104,7 @@ export async function kairo_deep_owned() {
 	destroy_owned();
 
 	return {
-		benchmark: 'kairo_deep_owned',
+		benchmark: 'kairo_triangle_owned',
 		time: timing.time.toFixed(2),
 		gc_time: timing.gc_time.toFixed(2)
 	};
