+++ benchmarking/benchmarks/ssr/index.js
@@ -1,0 +1,6 @@
+import { wrapper_bench } from "./wrapper/wrapper_bench.js";
+
+
+export const ssr_benchmarks = [
+	wrapper_bench
+]
